Login working with his form, fixed text for operation notifications, minor reestructuration of code and addition of comments in some archives
