+++ src/components/LoginUser/index.js
@@ -1,9 +1,12 @@
-import React, { useState } from "react"
+import React, { useEffect } from "react"
 import {FormikForm} from "../Forms"
 import {FormField} from "../Forms/formField"
 import { loginUserSchema } from "../Forms/schemas"
-import { FormContainer } from "components/CreateUser/styles"
-
+import { FormContainer } from './styles'
+import { useDispatch, useSelector} from 'react-redux'
+import { login,reset } from '../../store/slices/auth/index'
+import { alertToast } from 'services/alerts'
+import { useNavigate } from 'react-router-dom'
 const FormFields = () =>{
     return (
         <>
@@ -25,14 +28,35 @@ const FormFields = () =>{
 }
 
 export const LoginUserForm= () => {
-    const [values, setValues] = useState({
-        email: "",
-        password: ""
-    })
+    const dispatch = useDispatch()
+    const navigate = useNavigate()
+    const {user, isLoading, isError, isSuccess, message} = useSelector((state) => state.auth) 
+    const values = {
+        email: '',
+        password: ''
+      }
 
-    const handleSubmit= values=>{
-        console.log(values);
+    //Login req
+    const handleSubmit= values =>{
+        dispatch(login(values))
     }
+
+    //Effects/notifications 
+    useEffect(() => {
+        if(isError){
+            alertToast('error',message)
+        }
+        if(isSuccess || user){
+            alertToast('success','Inicio de sesión exitoso!')
+            navigate('/home')
+        }
+        dispatch(reset())
+    }, [user, isError, isSuccess, message, navigate, dispatch])
+
+    /*if(isLoading) {
+      //Loading screen, check preserve the state of the fields for not write all again after re-render
+    }*/
+
     return (
         <FormikForm
             title="Bienvenid@ a Somos Más"
